Version 0.0.1
Picorocket is in a runnable state. I wanted to achieve a Scene-Setup
that enables me to quickly change scenes without laborious initialzation
in each step.
In `main.lua` is the central game loop. There one can see how a
Scene-Table has to be structured.

The code from the pre-alpha was deleted, as it has found its way into
the corresponding scene-files.

There are some things to fixup or refactor, as

  - `Player` should be in ucfirst, as it is a global object
  - `PlanetsScene.planets`-Table is used in the `PlanetScene`, if so it
    should be a global table
  - The game has still to provide for any gameplay. Until now, only UI
    exists